make trait for shared fixtures/instantiations in processor tests
