+++ src/test/scala/org/constellation/ProcessorTest.scala
@@ -1,0 +1,53 @@
+package org.constellation
+
+import java.security.KeyPair
+
+import akka.actor.ActorSystem
+import akka.stream.ActorMaterializer
+import akka.testkit.TestProbe
+import org.constellation.Fixtures.{addPeerRequest, dummyTx, id}
+import org.constellation.crypto.KeyUtils
+import org.constellation.primitives.PeerData
+import org.constellation.primitives.Schema.{Id, Transaction}
+import org.constellation.util.APIClient
+import org.scalatest.AsyncFlatSpec
+
+import scala.concurrent.ExecutionContextExecutor
+
+trait ProcessorTest extends AsyncFlatSpec {
+  implicit val system: ActorSystem = ActorSystem("ProcessorTest")
+  implicit val materialize: ActorMaterializer = ActorMaterializer()
+//  implicit val executionContext: ExecutionContextExecutor = system.dispatcher
+
+  val keyPair: KeyPair = KeyUtils.makeKeyPair()
+  val peerData = PeerData(addPeerRequest, getAPIClient("", 1))
+  val peerManager = TestProbe()
+  val metricsManager = TestProbe()
+  val dbActor = TestProbe()
+  val mockData = new Data
+  mockData.updateKeyPair(keyPair)
+
+  val data = makeDao(mockData)
+  val tx: Transaction = dummyTx(data)
+  val invalidTx = dummyTx(data, -1L)
+  val srcHash = tx.src.hash
+  val txHash = tx.hash
+  val invalidSpendHash = invalidTx.hash
+  val randomPeer: (Id, PeerData) = (id, peerData)
+
+  def makeDao(mockData: Data, peerManager: TestProbe = peerManager, metricsManager: TestProbe = metricsManager,
+              dbActor: TestProbe = dbActor) = {
+    mockData.actorMaterializer = materialize
+    mockData.dbActor = dbActor.testActor
+    mockData.metricsManager = metricsManager.testActor
+    mockData.peerManager = peerManager.testActor
+    mockData
+  }
+
+  def getAPIClient(hostName: String, httpPort: Int) = {
+    val api = new APIClient().setConnection(host = hostName, port = httpPort)
+    api.id = id
+    api.udpPort = 16180
+    api
+  }
+}
